feat: move to react router 7 etc

+++ db/migrate.ts
@@ -1,15 +1,18 @@
-import { migrate } from 'drizzle-orm/bun-sqlite/migrator';
+import { migrate } from 'drizzle-orm/better-sqlite3/migrator';
 
-import { drizzle } from 'drizzle-orm/bun-sqlite';
+import { drizzle } from 'drizzle-orm/better-sqlite3';
 import * as schema from './schema.js';
-import { Database } from 'bun:sqlite';
+import Database from 'better-sqlite3';
 import { seed } from './seed.js';
 import fs from 'fs';
 import { setupFTS } from './fts.js';
 import { generateJson } from 'assets/html-to-json.js';
+import { sql } from 'drizzle-orm';
 
 // delete the database
-fs.unlinkSync('sqlite.db');
+if (fs.existsSync('sqlite.db')) {
+  fs.unlinkSync('sqlite.db');
+}
 const sqlite = new Database('sqlite.db');
 const db = drizzle(sqlite, {
   schema,
@@ -21,13 +24,20 @@ await migrate(db, {
 console.log('✅ Database migrated');
 
 console.log('Generating JSON');
-await generateJson();
+// await generateJson();
 console.log('✅ JSON generated');
 
 console.log('Seeding database');
-await seed(db);
-console.log('✅ Database seeded');
+try {
+  await seed(db);
+  console.log('✅ Database seeded');
+} catch (e) {
+  console.error(e);
+  process.exit(1);
+}
 
 console.log('Setting up FTS');
-await setupFTS(db);
+for (const statement of setupFTS(sql)) {
+  await db.run(statement);
+}
 console.log('✅ FTS setup');
